Fix browser.url regression when the previous request raised an exception.
If we had one successful request, then one request raising an exception the
browser.response was the response of the first request, causing a BrowserStateError
when trying to access browser.url.

Since the exceptions are passed into the test, the request was not successful and
we don't know the real url (we might have been redirected).
Therefore the url should be `None` in this situation.