Introduce request drivers.
Introduce a new concept of request drivers, which are responsible for
making requests and for responding to questions regarding the state and
the response.

The current implementation already supports ``Mechanize``, ``requests``
library and setting static HTML, resulting in three standard drivers.

The mechanism for switching the driver is the same as switching the
request library before this change.

The browser API did not change regarding tested public API.
Things such as accessing the last response directly with
``browser.response`` is no longer possible though.