Log exceptions to stderr.
When developing with the testbrowser a developer should see the
traceback of exceptions happening while publishing (e.g. in the view).
Because of the error handling only a 500 response was rendered.

We used to have logging enabled in the requests driver.
But this was only for this driver and it was noisy because it did log
other log entries too.

The new implementation is no longer in the driver but in the browser
core and it prints all entries added to Plone error_log while the
request is performed.
Exceptions are not printed within an expect_http_error context manager.