Add InsufficientPrivileges exception.
TheWhen the requests is sent to a Plone CMS and causes an
"insufficient privileges" result, a
``ftw.testbrowser.exceptions.InsufficientPrivileges` is raised.

The exception is raised for anonymous users (rendering the login form) as well
as for logged in users (rendering the "Insufficient Privileges" page).

The goal is to early abort and report problems, so that debugging
permission problems is easier.