Use current pages default `#form` in erroneous_fields.
Use the form identified by the id `#form` as the default form for
the erroneous_fields page object. For plone tests it can be assumed
that often the form under test is a content-type related form. Both
add and edit forms of plone content-types use the id `#form`. Thus we
provide this shortcut.