Make plone.protect import optional.
ftw.testbrowser should avoid dependencies on Plone packages, so that
it can be used outside of a Plone environment.

The send_authenticator added in ftw.testbrowser 1.25.0 introduced a
plone.protect import, causing the testbrowser to no longer work
without plone.protect.

By moving the plone.protect import from the module head to the code
send_authenticator code block, plone.protect is only imported when it
is actually used. The package using ftw.testbrowser with the
send_authenticator option is responsible to make sure that
plone.protect is correctly installed.