Provide a @nondefault_browsing wraper for page object tests.
Page objects should accept the browser instance as a keyword
argument.
They should make sure that all operations are executed with this
browser, recursively.
In order to test that, the @nondefault_browsing provides a way to run
the tests on a browser clone, so that the tests fail when page objects
access the default browser.