Plone 5: Process catalog queue for plone.protect auto support.
In order to support non-committing requests, changes may be made by an
earlier request or directly in the test code before executing a
request.
In order for the CSRF auto protection to not report false positives,
we need to mark all those changes as "safe" on the request right
before publishing the object, so that the CSRF protection only watches
the differences.

This concept does not work well with a lazy catalog queue, since the
work may not be done at the point where we marke objects as safe.
Therefore we must ensure that the queue is processed before we mark
object as safe.