Fix browser.debug to finish writing file before opening.
Fixes a bug in browser.debug, causing that the file is opened (e.g. in
the browser) before the data is flushed to the disk.

The reason for the bug is that the open subprocess is executed while
the filehandle is still open and therefore the written data may or may
not be written to the disk.