Respect HTTP response encoding in parser.
When a HTTP response contains only a partial of a HTML document, the
encoding declaration may not be included in the HTML.
In this case the lxml encoding detection may not choose the correct
enconding.
In order to improve parsing with the correct encoding, the HTML parser
is now configured to use the encoding provided in the content-type
HTTP response header.