Plone 5: disable resource registries (performance)
In Plone 5, the resource registries are cooked when the resource
registry viewlets are rendered.
Cooking the bundles takes a lot of time.
Since ``ftw.testbrowser`` does nothing with JavaScript or CSS, cooking
of resources is disabled by default for performance improvement.
That means that ``<script>`` and ``<styles>``
tags are missing in the HTML.
This can make the tests up to 4-5 times faster.